pyproject.toml
[tool.poetry]
name = "mybestpractices"
version = "0.2.0"
description = "Apply best practices to your Python code"
authors = ["Matthieu Gallet <github@19pouces.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
toml = "^0.10.2"
pyyaml = ">=5.1"
airspeed = "^0.5.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mybestpractices-copyright = "mybestpractices.copyright:main"
mybestpractices-git-autotag = "mybestpractices.git_autotag:main"
mybestpractices-structure = "mybestpractices.structure:main"
mybestpractices-third-party-update = "mybestpractices.vendor_update:main"
required-precommit-hooks = "mybestpractices.required_precommit_hooks:main"
